+++ src/algorithm/fork.rs
@@ -64,146 +64,18 @@ pub fn both(env: &mut Uiua) -> UiuaResult {
 pub fn fork(env: &mut Uiua) -> UiuaResult {
     let f = env.pop(FunctionArg(1))?;
     let g = env.pop(FunctionArg(2))?;
-    let a = env.pop(ArrayArg(1))?;
-    let b = env.pop(ArrayArg(2))?;
-
-    match g.signature().args {
-        0 | 1 => {
-            env.push(b.clone());
-            env.call(g)?;
-        }
-        2 => {
-            env.push(b.clone());
-            env.push(a.clone());
-            env.call(g)?;
-        }
-        n => {
-            return Err(env.error(format!(
-                "Fork's functions may not take more than 2 arguments, \
-                but function 2 one takes {n}"
-            )))
-        }
-    }
-
-    match f.signature().args {
-        0 | 1 => {
-            env.push(a);
-            env.call(f)?;
-        }
-        2 => {
-            env.push(b);
-            env.push(a);
-            env.call(f)?;
-        }
-        n => {
-            return Err(env.error(format!(
-                "Fork's functions may not take more than 2 arguments, \
-                but function 1 one takes {n}"
-            )))
-        }
-    }
-
-    Ok(())
-}
-
-pub fn trident(env: &mut Uiua) -> UiuaResult {
-    let f = env.pop(FunctionArg(1))?;
-    let g = env.pop(FunctionArg(2))?;
-    let h = env.pop(FunctionArg(3))?;
-    let a = env.pop(ArrayArg(1))?;
-    let b = env.pop(ArrayArg(2))?;
-    let c = env.pop(ArrayArg(3))?;
-
-    match h.signature().args {
-        0 | 1 => {
-            env.push(c.clone());
-            env.call(h)?;
-        }
-        2 => {
-            env.push(c.clone());
-            env.push(b.clone());
-            env.call(h)?;
-        }
-        3 => {
-            env.push(c.clone());
-            env.push(b.clone());
-            env.push(a.clone());
-            env.call(h)?;
-        }
-        n => {
-            return Err(env.error(format!(
-                "Trident's functions may not take more than 3 arguments, \
-                but the third function {h} takes {n}"
-            )))
-        }
-    }
-
-    match g.signature().args {
-        0 | 1 => {
-            env.push(b.clone());
-            env.call(g)?;
-        }
-        2 => {
-            env.push(c.clone());
-            env.push(a.clone());
-            env.call(g)?;
-        }
-        3 => {
-            env.push(c.clone());
-            env.push(b.clone());
-            env.push(a.clone());
-            env.call(g)?;
-        }
-        n => {
-            return Err(env.error(format!(
-                "Trident's functions may not take more than 3 arguments, \
-                but the second function {g} takes {n}"
-            )))
-        }
-    }
-
-    match f.signature().args {
-        0 | 1 => {
-            env.push(a);
-            env.call(f)?;
-        }
-        2 => {
-            env.push(b);
-            env.push(a);
-            env.call(f)?;
-        }
-        3 => {
-            env.push(c);
-            env.push(b);
-            env.push(a);
-            env.call(f)?;
-        }
-        n => {
-            return Err(env.error(format!(
-                "Trident's functions may not take more than 3 arguments, \
-                but the first function {f} takes {n}"
-            )))
-        }
-    }
-
-    Ok(())
-}
-
-pub fn share(env: &mut Uiua) -> UiuaResult {
-    let f = env.pop(FunctionArg(1))?;
-    let g = env.pop(FunctionArg(2))?;
     let f = f
         .into_func_array()
         .map_err(|val| {
             env.error(format!(
-                "Share' first function must be a function, but is has type {}",
+                "Fork's first function must be a function, but is has type {}",
                 val.type_name()
             ))
         })
         .and_then(|f| {
             f.into_scalar().map_err(|arr| {
                 env.error(format!(
-                    "Share' first function must be a scalar, but is has rank {}",
+                    "Fork's first function must be a scalar, but is has rank {}",
                     arr.rank()
                 ))
             })
@@ -212,14 +84,14 @@ pub fn share(env: &mut Uiua) -> UiuaResult {
         .into_func_array()
         .map_err(|val| {
             env.error(format!(
-                "Share' second function must be a function, but is has type {}",
+                "Fork's second function must be a function, but is has type {}",
                 val.type_name()
             ))
         })
         .and_then(|f| {
             f.into_scalar().map_err(|arr| {
                 env.error(format!(
-                    "Share' second function must be a scalar, but is has rank {}",
+                    "Fork's second function must be a scalar, but is has rank {}",
                     arr.rank()
                 ))
             })